requirements.txt
numpy>=1.23
pandas>=2.0
Pillow>=10.0
opencv-python>=4.8
tqdm>=4.65
kagglehub>=0.1